LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/unloading-documents/003-expected.txt

main frame - has 1 onunload handler(s)
frame "<!--frame1-->" - has 1 onunload handler(s)
CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.

FAIL document.open in beforeunload with link The quota has been exceeded.
